add transaction-signer docs, move AccountId to core
